biofm-mini/pyproject.toml
[tool.poetry]
name = "biofm-mini"
version = "0.1.0"
description = "Mini multimodal foundation model demo for biology with imaging, scRNA-seq, and clinical data"
authors = ["Artemis Folle <artemis@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "biofm", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.2.0"
torchvision = "^0.17.0"
numpy = "^1.26.0"
pandas = "^2.1.0"
scikit-learn = "^1.4.0"
pydantic = "^2.5.0"
hydra-core = "^1.3.2"
pyarrow = "^15.0.0"
tqdm = "^4.66.0"
typing-extensions = "^4.9.0"

[tool.poetry.extras]
scanpy = ["scanpy", "anndata"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"
black = "^24.2.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP", "S", "TID"]
ignore = ["B905", "E203"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --cov=biofm --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
no_implicit_optional = true
